Created validation rules for defining a type and source #3
Created a validation rule. This rule checks every logical source. First
it checks whether the logical source is part of a source group. If it is
part of a source group it checks whether a type is defined on the source
group and than it checks for the logical source. If for neither of them
a type is declared it will show an error. If for both of them a type is
declared it will show a warning.

I also did this validation rule for "source" since it the construction
is the same as for type.