related to issue #18
Added the table structure. Mapping is always created within the tables
array. Therefore when multiple mapping are created, still a valid csvw
is generated. I also removed the standard dialect property when no
dialect is declared. According to the CSVW best practice a dialect is
not mandatory on a mapping. Therefore it is only added when it is
declared.